infinite loop to download images when user icon is not available on server #50
Change the logic to below:
IconUrlUtil serves as an entry point for other code to call. (also contains some static methods to work with IconDownloadTask) Removed the old LruCache.
IconDownloadTask a place to divide task into 2 part: (it is just "download" from photo manager)
                    (1. Get bitmap from PhotoManager, and process the bitmap in background thread)
                    (2.set the processed bitmap to view)
PhotoManager:  an central place to fetch the bitmap. (it contains 3 level. Memory->Disk->network). When getting from network, it calls the static method instead of calling the IconDownload task again which will cause the infinite loop.